pyproject.toml
[project]
name = "autoslowmode"
version = "0.1.0"
description = "A Discord bot that automatically sets slowmode on channels based on the number of messages sent."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "hikari>=2.2.1",
    "hikari-arc[cron]>=2.0.0",
    "hikari-toolbox>=0.1.7",
    "python-dotenv>=1.1.0",
    "uvloop>=0.21.0 ; sys_platform == 'darwin' or sys_platform == 'linux'",
]
